requirements.txt
boto3==1.34.158
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.37.1
